+++ src/main/java/actions/ActionRegister.java
@@ -1,7 +1,7 @@
 package actions;
 
-import entities.AbstractEntity;
 import entities.AbstractPlayer;
+import entities.AbstractReactiveEntity;
 import java.io.Serializable;
 import static java.lang.System.out;
 import util.SafeList;
@@ -11,7 +11,7 @@ import util.SafeList;
  * It is used by the AbstractPlayer class to store Status effects and passives.
  */
 public class ActionRegister implements Serializable{
-	private final AbstractEntity registeredTo;
+	private final AbstractReactiveEntity registeredTo;
 	private final SafeList<OnHitListener> onHitRegister;
 	private final SafeList<OnHitListener> onBeHitRegister;
 	private final SafeList<OnHitListener> onMeleeHitRegister;
@@ -20,9 +20,9 @@ public class ActionRegister implements Serializable{
 	
     /**
      * Stores Listeners for an AbstractEntity.
-     * @param e the AbstractEntity to store Listeners for.
+     * @param e the AbstractReactiveEntity to store Listeners for.
      */
-	public ActionRegister(AbstractEntity e){
+	public ActionRegister(AbstractReactiveEntity e){
 		registeredTo = e;
 		onHitRegister = new SafeList<>();
 		onBeHitRegister = new SafeList<>();
